Fix comp warning

+++ src/deps/mpr/mprLib.c
@@ -11835,7 +11835,7 @@ int mprSamePath(MprCtx ctx, cchar *path1, cchar *path2)
         }
     } else {
         for (p1 = path1, p2 = path2; *p1 && *p2; p1++, p2++) {
-            if (tolower(*p1) != tolower(*p2) && !(isSep(fs, *p1) && isSep(fs, *p2))) {
+            if (tolower((int) *p1) != tolower((int) *p2) && !(isSep(fs, *p1) && isSep(fs, *p2))) {
                 break;
             }
         }
@@ -11878,7 +11878,7 @@ int mprSamePathCount(MprCtx ctx, cchar *path1, cchar *path2, int len)
         }
     } else {
         for (p1 = path1, p2 = path2; *p1 && *p2 && len > 0; p1++, p2++, len--) {
-            if (tolower(*p1) != tolower(*p2) && !(isSep(fs, *p1) && isSep(fs, *p2))) {
+            if (tolower((int) *p1) != tolower((int) *p2) && !(isSep(fs, *p1) && isSep(fs, *p2))) {
                 break;
             }
         }
